requirements.txt
PyPDF4
pytesseract
Pillow
pdf2image
cryptography
reportlab